rewrited middleware and settings logic
